+++ src/FluxFlow.Components.Timers/TimerComponentTypes.cs
@@ -7,4 +7,5 @@ public static class TimerComponentTypes
     public static readonly NodeType Delay = new("timer.delay");
     public static readonly NodeType Interval = new("timer.interval");
     public static readonly NodeType Schedule = new("timer.schedule");
+    public static readonly NodeType Throttle = new("timer.throttle");
 }