+++ src/FluxFlow.Components.Timers/TimerErrorCodes.cs
@@ -5,4 +5,5 @@ public static class TimerErrorCodes
     public const int DelayFailed = 9010;
     public const int IntervalFailed = 9000;
     public const int ScheduleFailed = 9020;
+    public const int ThrottleFailed = 9030;
 }